⚡ refactor: Apple Login Credential을 받는 부분을 ViewModel에서 ViewController가 받도록 리팩토링
